pyproject.toml
[project]
name = "ocr-hacking-bridge"
version = "0.1.0"
description = "Automated DARDE/DARDO document validation pipeline for ACH"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "Pillow>=10.0",
    "python-dateutil>=2.8",
    "httpx>=0.27",
    "streamlit>=1.30",
    "pymupdf>=1.24",
]
